requirements.txt
streamlit==0.64.0
lightgbm==3.2.1
pandas==1.1.2
numpy==1.19.2
seaborn==0.10.1
matplotlib==3.3.1
scipy==1.4.1
plotly==4.9.0
Pillow==8.2.0
scikit_learn==0.24.2
